+++ tests/fastapi/test_fastapi.py
@@ -36,10 +36,19 @@ def test_middleware_should_call_apilytics_api(
     }
 
     data = tests.conftest.decode_request_data(call_kwargs["data"])
-    assert data.keys() == {"path", "method", "statusCode", "timeMillis"}
+    assert data.keys() == {
+        "path",
+        "method",
+        "statusCode",
+        "requestSize",
+        "responseSize",
+        "timeMillis",
+    }
     assert data["path"] == "/"
     assert data["method"] == "GET"
     assert data["statusCode"] == 200
+    assert data["requestSize"] == 0
+    assert data["responseSize"] > 0
     assert isinstance(data["timeMillis"], int)
 
 
@@ -56,6 +65,59 @@ def test_middleware_should_send_query_params(
     assert data["path"] == "/dummy/123/path/"
     assert data["query"] == "param=foo&param2=bar"
     assert data["statusCode"] == 201
+    assert data["requestSize"] == 0
+    assert data["responseSize"] > 0
+    assert isinstance(data["timeMillis"], int)
+
+
+def test_middleware_should_send_zero_request_and_response_sizes(
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    response = client.post("/empty")
+    assert response.status_code == 200
+
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert data["requestSize"] == 0
+    assert data["responseSize"] == 0
+
+
+def test_middleware_should_send_non_zero_request_and_response_sizes(
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    response = client.post("/dummy?some=query", json={"hello": "world"})
+    assert response.status_code == 201
+
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert data["requestSize"] == 18
+    assert data["responseSize"] == 7  # `len(b"created")`
+
+
+def test_middleware_should_work_with_streaming_response(
+    mocked_urlopen: unittest.mock.MagicMock,
+) -> None:
+    response = client.get("/streaming")
+    assert response.status_code == 200
+
+    assert mocked_urlopen.call_count == 1
+    __, call_kwargs = mocked_urlopen.call_args
+    data = tests.conftest.decode_request_data(call_kwargs["data"])
+    assert data.keys() == {
+        "path",
+        "method",
+        "statusCode",
+        "requestSize",
+        "responseSize",
+        "timeMillis",
+    }
+    assert data["path"] == "/streaming"
+    assert data["method"] == "GET"
+    assert data["statusCode"] == 200
+    assert data["requestSize"] == 0
+    assert data["responseSize"] == 0  # Can't get body size from a streaming response.
     assert isinstance(data["timeMillis"], int)
 
 
@@ -88,7 +150,8 @@ def test_middleware_should_send_data_even_on_errors(
 
     __, call_kwargs = mocked_urlopen.call_args
     data = tests.conftest.decode_request_data(call_kwargs["data"])
-    assert data.keys() == {"method", "path", "timeMillis"}
+    assert data.keys() == {"method", "path", "timeMillis", "requestSize"}
     assert data["method"] == "GET"
     assert data["path"] == "/error"
+    assert data["requestSize"] == 0
     assert isinstance(data["timeMillis"], int)